Project/requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
image==1.5.33
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
django-import-export
